tasktracker/requirements.txt
cryptography==43.0.3
defusedxml==0.8.0rc2
pillow==11.0.0
pip==24.3.1
cffi==1.17.1
Django==4.2.16
asgiref==3.8.1
sqlparse==0.5.2
psycopg2-binary==2.9.10
requests==2.32.3
djangorestframework-simplejwt==5.3.1
djangorestframework==3.15.2
djoser==2.3.1
social-auth-core==4.5.4
social-auth-app-django==5.4.2
python3-openid==3.2.0
certifi==2024.8.30
urllib3==2.2.3
oauthlib==3.2.2
idna==3.10
charset-normalizer==3.4.0
pycparser==2.22
PyJWT==2.10.0
requests-oauthlib==2.0.0
django-filter==24.3